backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
psycopg[binary]
psycopg-pool
redis
requests
beautifulsoup4
lxml
yfinance
pandas
numpy
scikit-learn
joblib
python-dateutil
python-dotenv
